added error seperation for "Unkown function" vs "error in external function"

+++ vm.h
@@ -207,6 +207,6 @@ e_value e_create_array(e_vm* vm, e_value* arr, uint32_t arrlen, uint32_t index, 
 e_stack_status_ret e_api_stack_push(e_stack *stack, e_value v);
 e_stack_status_ret e_api_stack_pop(e_stack *stack);
 void e_api_register_sub(const char *identifier, uint32_t (*fptr)(e_vm *, uint32_t));
-uint8_t e_api_call_sub(e_vm *vm, const char *identifier, uint32_t arglen);
+int32_t e_api_call_sub(e_vm *vm, const char *identifier, uint32_t arglen);
 
 #endif //ES_VM_H
